+++ src/api_claude/utils.py
@@ -1,5 +1,5 @@
 import base64
-import config
+from models import app_settings
 from npy.core.logger import setup_logger
 
 logger = setup_logger()
@@ -18,7 +18,7 @@ def load_document_from_file(filepath: str) -> dict | None:
     logger.info(f"[api_claude] Load Document from filepath: {filepath}.")
     content_block: dict | None = None
 
-    matches = [(ext, doc_type) for ext, doc_type in config.AI_SUPPORTED_INPUT_FILETYPES.items()
+    matches = [(ext, doc_type) for ext, doc_type in app_settings.ai_supported_input_filetypes.items()
                if filepath.lower().endswith(ext)]
 
     if not matches: